cloud-functions/synthesize_long_audio/requirements.txt
functions-framework==3.*
google-cloud-texttospeech